potential fix

+++ src/server/StructSetItem.java
@@ -7,7 +7,7 @@ import java.util.Map;
 
 public class StructSetItem {
 
-    public byte completeCount, setItemID;
+    public int completeCount, setItemID;
     public Map<Integer, SetItem> items = new LinkedHashMap<Integer, SetItem>();
     public List<Integer> itemIDs = new ArrayList<Integer>();
